Refactor code for modularity
