+++ index.js
@@ -1,3 +1,5 @@
+import renderModule from './indexModule.js';
+
 class Book {
   constructor(Title, Author) {
     this.Title = Title;
@@ -12,34 +14,11 @@ const bookshelf = document.querySelector('.bookshelf');
 
 const awesomeBooks = localStorage.getItem('Books') ? JSON.parse(localStorage.getItem('Books')) : [];
 
-function createBookInfo(info) {
-  const pBook = document.createElement('p');
-  pBook.textContent = info;
-  return pBook;
-}
-
-function createButton(func, textContent, data, classes, index) {
-  const button = document.createElement('button');
-  button.type = func;
-  button.classList.add(`${classes}`);
-  button.setAttribute(`data-${data}`, index);
-  button.textContent = textContent;
-  return button;
-}
-
-function createBookDiv(title, author, index) {
-  const div = document.createElement('div');
-  div.setAttribute('data-book', `${index}`);
-  div.append(createBookInfo(title), createBookInfo(author));
-  div.appendChild(createButton('button', 'Remove', 'index', 'removeBtn', index));
-  return div;
-}
-
 // Displaying awesome books
 const displayBooks = () => {
   const bookshelfFrag = document.createDocumentFragment();
   awesomeBooks.forEach((book, i) => {
-    bookshelfFrag.appendChild(createBookDiv(book.Title, book.Author, i));
+    bookshelfFrag.appendChild(renderModule.createBookDiv(book.Title, book.Author, i));
   });
   bookshelf.appendChild(bookshelfFrag);
 };
@@ -48,9 +27,9 @@ const displayBooks = () => {
 const addBook = () => {
   const book = new Book(title.value.trim(), author.value.trim());
   const bookshelfArrLen = document.querySelectorAll('.bookshelf div').length;
-  if (title.value.length && author.value.length) { awesomeBooks.push(book); }
+  if (title.value.length && author.value.length) awesomeBooks.push(book);
   localStorage.setItem('Books', JSON.stringify(awesomeBooks));
-  bookshelf.appendChild(createBookDiv(title.value, author.value, bookshelfArrLen));
+  bookshelf.appendChild(renderModule.createBookDiv(title.value, author.value, bookshelfArrLen));
   title.value = '';
   author.value = '';
 };
